factor out InterruptChildren
