+++ sdk/go/common/util/cmdutil/child_js.go
@@ -38,3 +38,7 @@ func processExistsWithParent(pid int, ppid int) (bool, error) {
 func RegisterProcessGroup(cmd *exec.Cmd) {
 	panic(errors.New("unsupported"))
 }
+
+func InterruptChildren(pid int) error {
+	panic(errors.New("unsupported"))
+}
